Add multithreading operation for feature computation Add feature computation for single object
